+++ lib/features/menu/domain/model/contact.freezed.dart
@@ -1,0 +1,148 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'contact.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
+
+/// @nodoc
+mixin _$Contact {
+  String get name => throw _privateConstructorUsedError;
+  String? get phoneNumber => throw _privateConstructorUsedError;
+
+  @JsonKey(ignore: true)
+  $ContactCopyWith<Contact> get copyWith => throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $ContactCopyWith<$Res> {
+  factory $ContactCopyWith(Contact value, $Res Function(Contact) then) =
+      _$ContactCopyWithImpl<$Res, Contact>;
+  @useResult
+  $Res call({String name, String? phoneNumber});
+}
+
+/// @nodoc
+class _$ContactCopyWithImpl<$Res, $Val extends Contact>
+    implements $ContactCopyWith<$Res> {
+  _$ContactCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? name = null,
+    Object? phoneNumber = freezed,
+  }) {
+    return _then(_value.copyWith(
+      name: null == name
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as String,
+      phoneNumber: freezed == phoneNumber
+          ? _value.phoneNumber
+          : phoneNumber // ignore: cast_nullable_to_non_nullable
+              as String?,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$$ContactImplCopyWith<$Res> implements $ContactCopyWith<$Res> {
+  factory _$$$ContactImplCopyWith(
+          _$$ContactImpl value, $Res Function(_$$ContactImpl) then) =
+      __$$$ContactImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({String name, String? phoneNumber});
+}
+
+/// @nodoc
+class __$$$ContactImplCopyWithImpl<$Res>
+    extends _$ContactCopyWithImpl<$Res, _$$ContactImpl>
+    implements _$$$ContactImplCopyWith<$Res> {
+  __$$$ContactImplCopyWithImpl(
+      _$$ContactImpl _value, $Res Function(_$$ContactImpl) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? name = null,
+    Object? phoneNumber = freezed,
+  }) {
+    return _then(_$$ContactImpl(
+      name: null == name
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as String,
+      phoneNumber: freezed == phoneNumber
+          ? _value.phoneNumber
+          : phoneNumber // ignore: cast_nullable_to_non_nullable
+              as String?,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$$ContactImpl implements $Contact {
+  const _$$ContactImpl({required this.name, required this.phoneNumber});
+
+  @override
+  final String name;
+  @override
+  final String? phoneNumber;
+
+  @override
+  String toString() {
+    return 'Contact(name: $name, phoneNumber: $phoneNumber)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$$ContactImpl &&
+            (identical(other.name, name) || other.name == name) &&
+            (identical(other.phoneNumber, phoneNumber) ||
+                other.phoneNumber == phoneNumber));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, name, phoneNumber);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$$ContactImplCopyWith<_$$ContactImpl> get copyWith =>
+      __$$$ContactImplCopyWithImpl<_$$ContactImpl>(this, _$identity);
+}
+
+abstract class $Contact implements Contact {
+  const factory $Contact(
+      {required final String name,
+      required final String? phoneNumber}) = _$$ContactImpl;
+
+  @override
+  String get name;
+  @override
+  String? get phoneNumber;
+  @override
+  @JsonKey(ignore: true)
+  _$$$ContactImplCopyWith<_$$ContactImpl> get copyWith =>
+      throw _privateConstructorUsedError;
+}