+++ lib/services/di/services_dependency_container.dart
@@ -1,6 +1,8 @@
 import 'package:get_it/get_it.dart';
 import 'package:zachranobed/routes/app_router.dart';
 import 'package:zachranobed/services/auth_service.dart';
+import 'package:zachranobed/services/carrier_service.dart';
+import 'package:zachranobed/services/configuration_service.dart';
 import 'package:zachranobed/services/delivery_service.dart';
 import 'package:zachranobed/services/entity_pairs_service.dart';
 import 'package:zachranobed/services/entity_service.dart';
@@ -22,6 +24,8 @@ class ServicesDependencyContainer {
         GetIt.I<EntityPairService>(),
       ),
     );
+    GetIt.I.registerSingleton(ConfigurationService());
+    GetIt.I.registerSingleton(CarrierService());
     GetIt.I.registerSingleton(AppRouter());
   }
 }